refactor: removed other scheduled functions
