+++ src/index.js
@@ -69,10 +69,7 @@ const {
 } = require("./services/scheduledExecution");
 
 exports.wasteSchedPost = wasteSchedPost;
-exports.yearly_wasteSchedPost = yearlyWasteSchedPost;
-exports.monthlyStatusSchedPost = monthlyStatusSchedPost;
-exports.statusSchedPostDaily = statusSchedPostDaily;
-exports.weeklyStatusSchedPost = weeklyWasteSchedPost;
+// exports.statusSchedPostDaily = statusSchedPostDaily;
 
 app.listen(PORT, () => {
   console.log(`Listening at http://localhost:${PORT}`);
